project/requirements.txt
alembic==1.4.3
asgiref==3.2.10
boto3==1.16.0
botocore==1.19.6
click==7.1.2
Django==3.1.2
django-storages==1.10.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Migrate==2.5.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.4.1
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
